minor refactoring of messages type files
