--- x/slashrefund/types/message_claim.go
@@ -1,53 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-const TypeMsgClaim = "claim"
-
-var _ sdk.Msg = &MsgClaim{}
-
-func NewMsgClaim(delegatorAddress string, validatorAddress string) *MsgClaim {
-	return &MsgClaim{
-		DelegatorAddress: delegatorAddress,
-		ValidatorAddress: validatorAddress,
-	}
-}
-
-func (msg *MsgClaim) Route() string {
-	return RouterKey
-}
-
-func (msg *MsgClaim) Type() string {
-	return TypeMsgClaim
-}
-
-func (msg *MsgClaim) GetSigners() []sdk.AccAddress {
-	creator, err := sdk.AccAddressFromBech32(msg.DelegatorAddress)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{creator}
-}
-
-func (msg *MsgClaim) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
-func (msg *MsgClaim) ValidateBasic() error {
-
-	_, err := sdk.AccAddressFromBech32(msg.DelegatorAddress)
-	if err != nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid account address (%s)", err)
-	}
-
-	_, err = sdk.ValAddressFromBech32(msg.ValidatorAddress)
-	if err != nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid validator address (%s)", err)
-	}
-
-	return nil
-}
