--- x/slashrefund/types/message_deposit.go
@@ -1,58 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-const TypeMsgDeposit = "deposit"
-
-var _ sdk.Msg = &MsgDeposit{}
-
-func NewMsgDeposit(depositorAddress string, validatorAddress string, amount sdk.Coin) *MsgDeposit {
-	return &MsgDeposit{
-		DepositorAddress: depositorAddress,
-		ValidatorAddress: validatorAddress,
-		Amount:           amount,
-	}
-}
-
-func (msg *MsgDeposit) Route() string {
-	return RouterKey
-}
-
-func (msg *MsgDeposit) Type() string {
-	return TypeMsgDeposit
-}
-
-func (msg *MsgDeposit) GetSigners() []sdk.AccAddress {
-	depositor, err := sdk.AccAddressFromBech32(msg.DepositorAddress)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{depositor}
-}
-
-func (msg *MsgDeposit) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
-func (msg *MsgDeposit) ValidateBasic() error {
-
-	_, err := sdk.AccAddressFromBech32(msg.DepositorAddress)
-	if err != nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid depositor address (%s)", err)
-	}
-
-	_, err = sdk.ValAddressFromBech32(msg.ValidatorAddress)
-	if err != nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid validator address (%s)", err)
-	}
-
-	if !msg.Amount.IsValid() || msg.Amount.Amount.IsZero() {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "invalid denom or non-positive amount: (%s)", msg.Amount.String())
-	}
-
-	return nil
-}
