--- x/slashrefund/types/message_withdraw.go
@@ -1,58 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-const TypeMsgWithdraw = "withdraw"
-
-var _ sdk.Msg = &MsgWithdraw{}
-
-func NewMsgWithdraw(depositorAddress string, validatorAddress string, amount sdk.Coin) *MsgWithdraw {
-	return &MsgWithdraw{
-		DepositorAddress: depositorAddress,
-		ValidatorAddress: validatorAddress,
-		Amount:           amount,
-	}
-}
-
-func (msg *MsgWithdraw) Route() string {
-	return RouterKey
-}
-
-func (msg *MsgWithdraw) Type() string {
-	return TypeMsgWithdraw
-}
-
-func (msg *MsgWithdraw) GetSigners() []sdk.AccAddress {
-	creator, err := sdk.AccAddressFromBech32(msg.DepositorAddress)
-	if err != nil {
-		panic(err)
-	}
-	return []sdk.AccAddress{creator}
-}
-
-func (msg *MsgWithdraw) GetSignBytes() []byte {
-	bz := ModuleCdc.MustMarshalJSON(msg)
-	return sdk.MustSortJSON(bz)
-}
-
-func (msg *MsgWithdraw) ValidateBasic() error {
-
-	_, err := sdk.AccAddressFromBech32(msg.DepositorAddress)
-	if err != nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid depositor address (%s)", err)
-	}
-
-	_, err = sdk.ValAddressFromBech32(msg.ValidatorAddress)
-	if err != nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid validator address (%s)", err)
-	}
-
-	if !msg.Amount.IsValid() || msg.Amount.Amount.IsZero() {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "invalid denom or non-positive amount: (%s)", msg.Amount.String())
-	}
-
-	return nil
-}
